Update SketchFile.java
Sometimes, when i start the app, it crashes at line 521 due to an IndexOutOfBoundsException. The try-catch block prevent the app from crashing. I don't believe this is the best solution to that issue, but i don't know how the entire algorythm works. I only want to report the problem so you can solve it, probably better than me.